+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2016-06-14
 
 ### Added
+- recfun/BalanceSuite & recfun/PascalSuite: added extra test clauses.
 - recfun/Main.scala: findOpening and findClosing sub-routines added to balance() function.
 
 ## 2016-06-13